Fixed outline bug and bin bug
When binarizing an image, it used to be the case that limits would be fixed irrespective of user interaction with the limit boxes. 'ROI' outlines were bugged so they would only work if you were on your first interaction with the overlay map. ROI outlines would not consider the fact that clusters can be adjacent, creating outlines that are on top of each other.